+++ magda-gateway/src/createGenericProxyRouter.ts
@@ -2,12 +2,12 @@ import express from "express";
 import { Router } from "express";
 import urijs from "urijs";
 import escapeStringRegexp from "escape-string-regexp";
-
-import buildJwt from "magda-typescript-common/src/session/buildJwt";
-
 import createBaseProxy from "./createBaseProxy";
 import Authenticator from "./Authenticator";
 import { TenantMode } from "./setupTenantMode";
+import buildJwtFromReq from "magda-typescript-common/src/session/buildJwtFromReq";
+import createApiAccessControlMiddleware from "./createApiAccessControlMiddleware";
+import AuthDecisionQueryClient from "magda-typescript-common/src/opa/AuthDecisionQueryClient";
 
 export type ProxyTarget = DetailedProxyTarget | string;
 export type MethodWithProxyTaget = {
@@ -19,6 +19,7 @@ export interface DetailedProxyTarget {
     to: string;
     methods?: ProxyMethodType[];
     auth?: boolean;
+    accessControl?: boolean;
     redirectTrailingSlash?: boolean;
     statusCheck?: boolean;
 }
@@ -32,6 +33,7 @@ export interface GenericProxyRouterOptions {
     tenantMode: TenantMode;
     defaultCacheControl?: string;
     proxyTimeout?: number;
+    authClient: AuthDecisionQueryClient;
 }
 
 /**
@@ -68,6 +70,7 @@ export default function createGenericProxyRouter(
 
     const authenticator = options.authenticator;
     const jwtSecret = options.jwtSecret;
+    const authClient = options.authClient;
 
     const router: Router = express.Router();
 
@@ -75,7 +78,7 @@ export default function createGenericProxyRouter(
         if (jwtSecret && req.user) {
             proxyReq.setHeader(
                 "X-Magda-Session",
-                buildJwt(jwtSecret, req.user.id, { session: req.user.session })
+                buildJwtFromReq(req, jwtSecret)
             );
         }
     });
@@ -85,12 +88,24 @@ export default function createGenericProxyRouter(
         target: string,
         verbs: ProxyMethodType[] = ["all"],
         auth = false,
-        redirectTrailingSlash = false
+        redirectTrailingSlash = false,
+        accessControl = false
     ) {
-        console.log("PROXY", baseRoute, target, verbs);
+        console.log(
+            "PROXY",
+            baseRoute,
+            target,
+            verbs,
+            "auth:",
+            auth,
+            "accessControl: ",
+            accessControl,
+            "redirectTrailingSlash: ",
+            redirectTrailingSlash
+        );
         const routeRouter: any = express.Router();
 
-        if (authenticator && auth) {
+        if (authenticator && (auth || accessControl)) {
             authenticator.applyToRoute(routeRouter);
         }
 
@@ -98,6 +113,12 @@ export default function createGenericProxyRouter(
             if (typeof verb === "string") {
                 routeRouter[verb.toLowerCase()](
                     "*",
+                    createApiAccessControlMiddleware(
+                        authClient,
+                        baseRoute,
+                        jwtSecret,
+                        accessControl
+                    ),
                     (req: express.Request, res: express.Response) => {
                         proxy.web(req, res, { target });
                     }
@@ -113,6 +134,12 @@ export default function createGenericProxyRouter(
                     typeof verb?.target === "string" ? verb.target : target;
                 routeRouter[method](
                     "*",
+                    createApiAccessControlMiddleware(
+                        authClient,
+                        baseRoute,
+                        jwtSecret,
+                        accessControl
+                    ),
                     (req: express.Request, res: express.Response) => {
                         proxy.web(req, res, { target: runtimeTarget });
                     }
@@ -163,8 +190,9 @@ export default function createGenericProxyRouter(
                 path,
                 target.to,
                 target.methods,
-                !!target.auth,
-                target.redirectTrailingSlash
+                !!target?.auth,
+                target.redirectTrailingSlash,
+                !!target?.accessControl
             );
         });