+++ magda-registry-api/src/main/scala/au/csiro/data61/magda/registry/RecordsService.scala
@@ -36,7 +36,10 @@ import scala.concurrent.{Await, ExecutionContext, Future}
 import scala.concurrent.duration._
 import scala.util.{Failure, Success}
 
-import au.csiro.data61.magda.directives.CommonDirectives.withBlockingTask
+import au.csiro.data61.magda.directives.CommonDirectives.{
+  onCompleteBlockingTask,
+  onCompleteBlockingTaskIn
+}
 
 @Path("/records")
 @io.swagger.annotations.Api(value = "records", produces = "application/json")
@@ -155,7 +158,7 @@ class RecordsService(
           )
         ) {
           requiresSpecifiedTenantId { tenantId =>
-            withBlockingTask {
+            onCompleteBlockingTask {
               val theResult = DB localTx { implicit session =>
                 session.queryTimeout(this.defaultQueryTimeout)
                 recordPersistence
@@ -271,10 +274,7 @@ class RecordsService(
           requiresSpecifiedTenantId { tenantId =>
             parameters('sourceTagToPreserve, 'sourceId) {
               (sourceTagToPreserve, sourceId) =>
-                extractActorSystem { routeActor =>
-                  implicit val blockingExeCtx = routeActor.dispatchers.lookup(
-                    "long-running-db-operation-dispatcher"
-                  )
+                onCompleteBlockingTaskIn("long-running-db-operation-dispatcher") {
                   val deleteFuture = Future {
                     // --- DB session needs to be created within the `Future`
                     // --- as the `Future` will keep running after timeout and require active DB session
@@ -322,6 +322,7 @@ class RecordsService(
                         )
                       )
                   }
+
                 }
             }
           }
@@ -427,7 +428,7 @@ class RecordsService(
                 Left(recordIn),
                 merge.getOrElse(false)
               ) {
-                withBlockingTask {
+                onCompleteBlockingTask {
                   val result = DB localTx { implicit session =>
                     session.queryTimeout(this.defaultQueryTimeout)
                     recordPersistence.putRecordById(
@@ -582,7 +583,7 @@ class RecordsService(
               id,
               Right(recordPatch)
             ) {
-              withBlockingTask {
+              onCompleteBlockingTask {
                 val theResult = DB localTx { implicit session =>
                   session.queryTimeout(this.defaultQueryTimeout)
                   recordPersistence.patchRecordById(
@@ -711,7 +712,7 @@ class RecordsService(
             authClient,
             AuthDecisionReqConfig("object/record/update")
           ) { authDecision =>
-            withBlockingTask {
+            onCompleteBlockingTask {
               val result = DB localTx { implicit session =>
                 session.queryTimeout(this.defaultQueryTimeout)
                 recordPersistence.patchRecords(
@@ -838,7 +839,7 @@ class RecordsService(
                 authClient,
                 AuthDecisionReqConfig("object/record/update")
               ) { authDecision =>
-                withBlockingTask {
+                onCompleteBlockingTask {
                   val result = DB localTx { implicit session =>
                     session.queryTimeout(this.defaultQueryTimeout)
                     recordPersistence.putRecordsAspectById(
@@ -961,7 +962,7 @@ class RecordsService(
               authClient,
               AuthDecisionReqConfig("object/record/update")
             ) { authDecision =>
-              withBlockingTask {
+              onCompleteBlockingTask {
                 val result = DB localTx { implicit session =>
                   session.queryTimeout(this.defaultQueryTimeout)
                   recordPersistence.deleteRecordsAspectArrayItems(
@@ -1089,7 +1090,7 @@ class RecordsService(
                 )
               )
             ) {
-              withBlockingTask {
+              onCompleteBlockingTask {
                 val result = DB localTx { implicit session =>
                   session.queryTimeout(this.defaultQueryTimeout)
                   recordPersistence