+++ magda-registry-api/src/main/scala/au/csiro/data61/magda/registry/RecordsServiceRO.scala
@@ -19,7 +19,7 @@ import au.csiro.data61.magda.client.{AuthDecisionReqConfig}
 import au.csiro.data61.magda.model.AspectQuery
 import scalikejdbc.interpolation.SQLSyntax
 import au.csiro.data61.magda.directives.RouteDirectives.completeBlockingTask
-import au.csiro.data61.magda.directives.CommonDirectives.withBlockingTask
+import au.csiro.data61.magda.directives.CommonDirectives.onCompleteBlockingTask
 
 import scala.concurrent.ExecutionContext
 
@@ -356,7 +356,7 @@ class RecordsServiceRO(
                 )
               } else {
                 completeBlockingTask {
-                  DB localTx { implicit session =>
+                  DB readOnly { implicit session =>
                     session.queryTimeout(this.defaultQueryTimeout)
                     recordPersistence.getAllWithAspects(
                       tenantId,
@@ -934,7 +934,7 @@ class RecordsServiceRO(
         ) { authDecision =>
           parameters('aspect.*, 'optionalAspect.*, 'dereference.as[Boolean].?) {
             (aspects, optionalAspects, dereference) =>
-              withBlockingTask {
+              onCompleteBlockingTask {
                 DB readOnly { implicit session =>
                   session.queryTimeout(this.defaultQueryTimeout)
                   recordPersistence.getByIdWithAspects(
@@ -1031,7 +1031,7 @@ class RecordsServiceRO(
           authApiClient,
           AuthDecisionReqConfig("object/record/read")
         ) { authDecision =>
-          withBlockingTask {
+          onCompleteBlockingTask {
             DB readOnly { implicit session =>
               session.queryTimeout(this.defaultQueryTimeout)
               recordPersistence
@@ -1118,7 +1118,7 @@ class RecordsServiceRO(
           authApiClient,
           AuthDecisionReqConfig("object/record/read")
         ) { authDecision =>
-          withBlockingTask {
+          onCompleteBlockingTask {
             DB readOnly { implicit session =>
               session.queryTimeout(this.defaultQueryTimeout)
               recordPersistence