+++ magda-scala-common/src/main/scala/au/csiro/data61/magda/util/SimpleTimer.scala
@@ -1,0 +1,9 @@
+package au.csiro.data61.magda.util
+
+class SimpleTimer {
+  val startTime = System.currentTimeMillis()
+
+  def elapsed() = {
+    System.currentTimeMillis() - startTime
+  }
+}