+++ magda-web-client/src/actions/datasetSearchActions.ts
@@ -2,13 +2,14 @@ import buildSearchQueryString, {
     Query
 } from "../helpers/buildSearchQueryString";
 import { actionTypes } from "../constants/ActionTypes";
-import { FetchError, Dispatch } from "../types";
+import { FetchError } from "../types";
 import {
     DataSearchJson,
     FacetAction,
     SearchAction
 } from "../helpers/datasetSearch";
 import { searchDatasets } from "api-clients/SearchApis";
+import { retrieveLocalData, setLocalData } from "storage/localStorage";
 
 export function requestResults(
     queryObject: any,
@@ -46,10 +47,22 @@ export function resetDatasetSearch(): SearchAction {
 }
 
 export function fetchSearchResults(queryObject: Query) {
-    return (dispatch: Dispatch) => {
+    return (dispatch, getState) => {
+        const state = getState();
+        const magdaSearchDatasetUserId = retrieveLocalData(
+            "magdaSearchDatasetUserId",
+            ""
+        );
+        const currentUserId = state?.userManagement?.user?.id
+            ? state.userManagement.user.id
+            : "";
+        const noCache = magdaSearchDatasetUserId !== currentUserId;
+        if (noCache) {
+            setLocalData("magdaSearchDatasetUserId", currentUserId);
+        }
         const queryString = buildSearchQueryString(queryObject);
         dispatch(requestResults(queryObject, queryString));
-        searchDatasets(queryObject)
+        searchDatasets(queryObject, noCache)
             .then((json) => dispatch(receiveResults(queryString, json)))
             .catch((error) =>
                 dispatch(